replace augmented graph with complete graph with edge costs that vary based on input
The primary reason for this is to enable us to use contraction hierarchies, which
require the structure of the graph to not change between queries.

This also results in a performance improvement for queries which allow cars, while
hurting queries that disallow cars. Since we expect queries allowing cars to be
more numerous, this still seems like a worthwhile tradeoff.