parallelize fastest route computation
We add each route query to a thread pool to perform the work off of the main thread. On my laptop
with 16 cores, this yields a roughly 25% performance improvement over the previous single-threaded
implementation on //engine/tests:simulation_timer for 7 days. This is close to the theoretical
maximum gain from parallelizing route calculation since previous benchmarking showed that ~30% of
total compute was spent calculating fastest routes.