calculate and display isochrone maps
https://en.wikipedia.org/wiki/Isochrone_map

Uses a weighted distance transform, patched from a standard distance transform in the imageproc
crate, to produce the final maps. The actual travel times are computed naively; if we want to
compute isochrone maps as part of the game loop, or if we want to compute lots of isochrone maps
simultaneously, we will want to patch fast_paths to support this.