railway/network refactor
Previously metro lines were their own entity, and had no notion of shared tracks. We now have a
separate railway system, which uses the network system (also used by highways). Metro lines are
built on top of the railways. In this way, we correctly represent multiple metro lines sharing the
same track. Critically, this will make it possible to edit the networks in an intuitive way.

As part of this refactor, we also change the generation code. The network code is shared by
highways and railways. This should also be a much more robust system going forward.

The tests are passing, but many things are not yet fully working: for exampel, maps other than SF,
timetable_tool, and properu metro line timing.